slightly amended api structure
